Created Burst Projectile Character
Shoots x amount of projectiles in an interval of x seconds. Also kind of had to cheat and make it go into the PlayerAttack script to actually use the weapon since the burst timing needed to be done in a Monobehavior script instead of the CharacterSO class. This shouldn't be a problem as it's the only one that needs this sort of work around.