Added bowling ball character
Also added a EntityBorder object that has an EntityDestroyer script that basically destroys everything it touches. This is for the bowling ball and any other object that isn't programmed to do anything after it falls. The bowling ball can only be used when the original bowling ball is gone, so that means it must roll off the entire map for the player to throw another bowling ball