Everything works... KINDA
The List in the editor causes major breakage for reasons that I don't understand. I'm assuming it's just a Unity bug since the List freaks out everytime I touch it so as long as the MANAGERS object isn't being inspected while playing, things should work. Also is only in the SampleScene for right now